Automatic time table generation
In early days, timetable scheduling was done manually with a single person or some group involveld in taskof scheduling was their hands, which take lot of effort and time. While scheduling even the smallest constraints can take a lot of time and the case is even worse when the number of constraints or the amounts of data to dal with increase .Institutions/Schools/Collages/University are the regular uses of such timetables. They need to schedule their course to meet the need of current duration and facilities that are available to them.

our aim here is to develop a simple , easily understandable ,efficient and portable application, which could automatically generate good quality timetable within seconds.